Trying to move pagination code out to base.py, but errors...
flask_classy.DecoratorCompatibilityError - will need to refactor,
maybe into functions
